+++ boussole/__init__.py
@@ -1,2 +1,2 @@
 """Commandline interface to build Sass projects using libsass-python"""
-__version__ = '1.2.3-pre.1'
+__version__ = '1.2.3-pre.2'