+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Version 1.2.3 - Unreleased
 --------------------------
 
 * **Introduced new settings** ``CUSTOM_IMPORT_EXTENSIONS`` which default value is ``['.css']`` to keep CSS source including behavior as default just like before libsass==3.5.3, close #29;
+* Fixed source map url, close #28;
 
 Version 1.2.2 - 2017/12/12
 --------------------------
@@ -30,7 +31,7 @@ Version 1.2.0 - 2017/01/21
 * Updated tests requirements;
 * Removed python 3.4 from tox envs;
 * Added ``--poll`` option on watch command to use Watchdog polling observer instead of the native platform observer, close #22;
-* Fixed compiler tests for changes about sourcemap since last libsass version;
+* Fixed compiler tests for changes about source map since last libsass version;
 * Fixed Sass brand name according to http://sassnotsass.com/;
 * Validated tests with ``libsass==0.12.3``;